greystone/requirements.txt
Django==1.8.4
django-autoslug==1.7.2
django-filter==0.14.0
djangorestframework==3.4.6
docutils==0.12
drf-generators==0.2.8
GitPython==1.0.1
python-dateutil==2.4.2
pytz==2016.6.1
requests==2.6.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1

